chore(engine,maker): migrate tests vitest → @gjsify/unit
Completes the workspace-wide migration started by the
signalling-server. Tests now run on BOTH Node + GJS via
`gjsify test`.

Why @gjsify/unit:

  GJS-only modules (LanPublisher/Browser via Gio.Subprocess,
  the eventual @gjsify/webrtc / @gjsify/ws integration) become
  testable in their target runtime instead of being shadowed
  by Node-only mocks. Consistent with the rest of the gjsify-
  native pipeline; one fewer external dev dep.

Mapping cheat sheet:

  Import:
    `import {describe, it, expect, vi} from 'vitest'`
    →`import {describe, it, expect, spy} from '@gjsify/unit'`

  Wrapping:
    Each test file's tests now sit under
    `export default async () => { await describe(...) }`
    A `src/test.mts` per package imports + `run()`s all suites.

  Equality:
    @gjsify/unit's `toEqual` is `==` (loose) — for vitest's deep
    semantics use `toStrictEqual`. `toMatchObject` has no direct
    equivalent; replaced with explicit field assertions.

  Spies:
    `vi.fn()` → `spy(fnStub)` (the no-arg `spy()` overload
    defaults to `Spy<() => void>`, so pass a typed stub for
    real signatures).
    `vi.spyOn(obj, 'm').mockImplementation(...)` → `obj.m = ...`
    inline replacement (the codebase only used it to silence
    `console.warn`; replaced with a `muteWarn(fn)` helper).

  Lifecycle:
    `beforeAll`/`afterAll` (used only by the signalling-server
    e2e) → top-level setup + `try { describe } finally
    { teardown }` inside `on('Node.js', …)`.

Engine specifics:

  vitest.setup.ts polyfilled `globalThis.window` so excalibur's
  import-time `polyfill()` didn't throw under Node. The same
  polyfill now sits at the top of `src/test.mts` (runs on Node;
  no-op on GJS where window is undefined).

Package.json `build:barrels`:

  gjsify barrels would otherwise include `./test.mts` in the
  generated index.ts barrel, which tsc can't resolve. Explicit
  `--exclude '^test\\.mts$'` (plus restoring the default
  `.spec.` / `.test.` excludes that the override clobbers).
  Applied to signalling-server + maker-gjs (engine ships a
  hand-written index.ts).

Per-package devDeps:

  - vitest dropped
  - @gjsify/unit ^0.4.32 added

Verified end-to-end on this machine:

  gjsify workspace @pixelrpg/signalling-server test
    → 26 expects  ✅ gjs ✅ node

  gjsify workspace @pixelrpg/engine test
    → 166 expects  ✅ gjs ✅ node

  gjsify workspace @pixelrpg/maker-gjs test
    → 20 expects  ✅ gjs ✅ node

  gjsify foreach check + build → clean

Total: 212 cross-runtime test calls.

Test files renamed `*.test.ts` → `*.spec.ts` to match the
@gjsify/unit convention (the file-suffix convention is the
discriminator the gjsify-native build infrastructure assumes).
