+++ apps/maker-gjs/src/test.mts
@@ -1,0 +1,6 @@
+import { run } from '@gjsify/unit'
+
+import lanDiscoveryParseSuite from './services/lan-discovery-parse.spec.js'
+import lanSignallingSuite from './services/lan-signalling.spec.js'
+
+run({ lanDiscoveryParseSuite, lanSignallingSuite })
