+++ apps/signalling-server/package.json
@@ -7,8 +7,8 @@
   "private": true,
   "scripts": {
     "clear": "rm -rf dist org.pixelrpg.signalling-server",
-    "build:barrels": "gjsify barrels src",
-    "check:barrels": "gjsify barrels --check src",
+    "build:barrels": "gjsify barrels src --exclude '\\.spec\\.' --exclude '\\.test\\.' --exclude '\\.test-data\\.' --exclude '^test\\.mts$'",
+    "check:barrels": "gjsify barrels --check src --exclude '\\.spec\\.' --exclude '\\.test\\.' --exclude '\\.test-data\\.' --exclude '^test\\.mts$'",
     "build:app": "gjsify build src/main.ts --app gjs --globals auto --shebang --outfile org.pixelrpg.signalling-server",
     "build": "gjsify run build:barrels && gjsify run build:app",
     "check": "gjsify run build:barrels && tsc --noEmit",
