--- packages/engine/vitest.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    include: ['src/**/*.test.ts'],
-    environment: 'node',
-    setupFiles: ['./vitest.setup.ts'],
-  },
-})
